Added in a useModal component to reduce code duplication on pages that are using the modal.
