+++ src/components/dictionary-page/word-card-desktop.tsx
@@ -1,36 +1,18 @@
-import React, { useEffect } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 import WordModal from './word-modal';
 import { FvWord } from '../common/data';
 import Modal from '../common/modal/modal';
+import { useModal } from '../common/use-modal/use-modal';
 
 function WordCardDesktop({ term }: FvWord) {
   const location = useLocation();
-  const [showModal, setShowModal] = React.useState(
+  const { setShowModal, showModal, closeModal } = useModal(
     (location.hash === `#${term.source}-${term.entryID}` ||
       location.hash === `#${term.source}-${term.entryID}?source=/profile`) &&
       window.matchMedia('(min-width: 768px').matches
   );
+
   const { word, definition, audio } = term;
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    if (showModal) {
-      document.body.style.overflow = 'hidden';
-    }
-    return () => {
-      document.body.style.overflow = 'unset';
-    };
-  }, [showModal]);
-
-  function closeModal() {
-    setShowModal(false);
-    const sourcePageUrl = window.location.hash.split('?')[1]?.split('=')[1];
-
-    if (sourcePageUrl) {
-      navigate(sourcePageUrl);
-    }
-  }
 
   return (
     <>
